requirements.txt
streamlit==1.28.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
numpy==1.24.3
transformers==4.30.0  # For Hugging Face models
easyocr==1.6.2        # For text extraction